+++ test/api/auth-api-test.js
@@ -1,26 +1,26 @@
 import { assert } from "chai";
 import { lighthouseService } from "./lighthouse-service.js";
 import { decodeToken } from "../../src/api/jwt-utils.js";
-import { maggie } from "../fixtures.js";
+import { maggie, maggieCredentials } from "../fixtures.js";
 
 suite("Authentication API tests", async () => {
   setup(async () => {
     lighthouseService.clearAuth();
     await lighthouseService.createUser(maggie);
-    await lighthouseService.authenticate(maggie);
+    await lighthouseService.authenticate(maggieCredentials);
     await lighthouseService.deleteAllUsers();
   });
 
   test("authenticate", async () => {
     const returnedUser = await lighthouseService.createUser(maggie);
-    const response = await lighthouseService.authenticate(maggie);
+    const response = await lighthouseService.authenticate(maggieCredentials);
     assert(response.success);
     assert.isDefined(response.token);
   });
 
   test("verify Token", async () => {
     const returnedUser = await lighthouseService.createUser(maggie);
-    const response = await lighthouseService.authenticate(maggie);
+    const response = await lighthouseService.authenticate(maggieCredentials);
 
     const userInfo = decodeToken(response.token);
     assert.equal(userInfo.email, returnedUser.email);