+++ test/api/lighthouse-service.js
@@ -8,6 +8,7 @@ export const lighthouseService = {
   async authenticate(user) {
     const response = await axios.post(`${this.lighthouseUrl}/api/users/authenticate`, user);
     axios.defaults.headers.common.Authorization = `Bearer ${  response.data.token}`;
+    console.log("Token" ,response.data.token);
     return response.data;
   },